# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.sendDocumentMessageBody:
    properties:
      base64:
        type: string
      filename:
        type: string
      phone:
        type: string
    type: object
  handler.sendDocumentMessageResponse:
    properties:
      message:
        $ref: '#/definitions/response.Message'
    type: object
  handler.sendImageMessageBody:
    properties:
      base64:
        type: string
      phone:
        type: string
    type: object
  handler.sendImageMessageResponse:
    properties:
      message:
        $ref: '#/definitions/response.Message'
    type: object
  handler.sendTextMessageBody:
    properties:
      phone:
        type: string
      text:
        type: string
    type: object
  handler.sendTextMessageResponse:
    properties:
      message:
        $ref: '#/definitions/response.Message'
    type: object
  response.Message:
    properties:
      body:
        type: string
      chat:
        type: string
      from_me:
        type: boolean
      id:
        type: integer
      media_base64:
        type: string
      media_mimetype:
        type: string
      media_type:
        type: string
      message_id:
        type: string
      sender:
        type: string
      timestamp:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API to handle multiple WhatsApp instances
  title: ZapMeow API
  version: "1.0"
paths:
...
  /{instanceId}/chat/send/document:
    post:
      consumes:
      - application/json
      description: Sends an Document message on WhatsApp using the specified instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Document message body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.sendDocumentMessageBody'
      produces:
      - application/json
      responses:
        "200":
          description: Message Send Response
          schema:
            $ref: '#/definitions/handler.sendDocumentMessageResponse'
      summary: Send Document Message on WhatsApp
      tags:
      - WhatsApp Chat
  /{instanceId}/chat/send/image:
    post:
      consumes:
      - application/json
      description: Sends an image message on WhatsApp using the specified instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Image message body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.sendImageMessageBody'
      produces:
      - application/json
      responses:
        "200":
          description: Message Send Response
          schema:
            $ref: '#/definitions/handler.sendImageMessageResponse'
      summary: Send Image Message on WhatsApp
      tags:
      - WhatsApp Chat
  /{instanceId}/chat/send/text:
    post:
      consumes:
      - application/json
      description: Sends a text message on WhatsApp using the specified instance.
      parameters:
      - description: Instance ID
        in: path
        name: instanceId
        required: true
        type: string
      - description: Text message body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.sendTextMessageBody'
      produces:
      - application/json
      responses:
        "200":
          description: Message Send Response
          schema:
            $ref: '#/definitions/handler.sendTextMessageResponse'
      summary: Send Text Message on WhatsApp
      tags:
      - WhatsApp Chat
...
swagger: "2.0"